refactor: UserRoleEnum 삭제
-공통모듈에 포함된 enum사용
