+++ com.fn.ai.user/src/main/java/com/fn/ai/user/model/UserRespository.java
@@ -1,11 +1,10 @@
 package com.fn.ai.user.model;
 
-import com.fn.ai.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRespository extends JpaRepository<User, UUID> {
-    Optional<User> findByUsername(String username);
+
+  Optional<User> findByUsername(String username);
 }
