+++ com.fn.ai.user/src/main/java/com/fn/ai/user/presentation/dto/UserSignInResponseDto.java
@@ -1,15 +1,14 @@
 package com.fn.ai.user.presentation.dto;
 
-import com.fn.ai.user.UserRoleEnum;
-import lombok.Builder;
-
+import com.fn.ai.common.context.UserRoleEnum;
 import java.util.UUID;
+import lombok.Builder;
 
 @Builder
 public record UserSignInResponseDto(UUID id,
                                     String username,
                                     String password,
-                                    UserRoleEnum role){
+                                    UserRoleEnum role) {
 
 }
